mobile_api/requirements.txt
fastapi
uvicorn
requests
google-cloud-bigquery
google-auth
python-dotenv
pytz